requirements.txt
pandas 
schedule 
speedtest-cli
